+++ alacritty_terminal/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Pass `-q` to `login` on macOS if `~/.hushlogin` is present
+- Use `pub` visibility instead of `pub(crate)` for `Shell::program` and `Shell::args`
 
 ## 0.25.0